Move all GraphQL code to lib
