+++ decidim-api/lib/decidim/api.rb
@@ -5,13 +5,8 @@ module Decidim
   # This module holds all business logic related to exposing a Public API for
   # decidim.
   module Api
-    autoload :MutationType, "decidim/api/types/mutation"
-    autoload :QueryType, "decidim/api/types/query"
-    autoload :AuthorInterface, "decidim/api/types/author_interface"
-
-    autoload :TranslatedFieldType, "decidim/api/types/translated_field"
-    autoload :LocalizedStringType, "decidim/api/types/localized_string"
-
+    autoload :MutationType, "decidim/api/mutation_type"
+    autoload :QueryType, "decidim/api/query_type"
     autoload :Schema, "decidim/api/schema"
   end
 end
