+++ decidim-api/lib/decidim/api/query_type.rb
