+++ decidim-comments/lib/decidim/comments.rb
@@ -7,5 +7,8 @@ module Decidim
   # a React component which handle all the comments render and logic.
   module Comments
     autoload :CommentsHelper, "decidim/comments/comments_helper"
+    autoload :AddCommentType, "decidim/comments/api/add_comment_type"
+    autoload :CommentMutationType, "decidim/comments/api/comment_mutation_type"
+    autoload :CommentType, "decidim/comments/api/comment_type"
   end
 end
