+++ decidim-comments/lib/decidim/comments/api/add_comment_type.rb
