+++ decidim-core/lib/decidim/core.rb
@@ -2,6 +2,7 @@
 # frozen_string_literal: true
 require "decidim/core/engine"
 require "decidim/core/version"
+require "decidim/core/api"
 
 # Decidim configuration.
 module Decidim
