+++ decidim-core/lib/decidim/core/api/process_type.rb
@@ -7,7 +7,7 @@ module Decidim
 
     field :id, !types.ID, "The Process' unique ID"
 
-    field :title, Api::TranslatedFieldType, "The title of this process."
+    field :title, TranslatedFieldType, "The title of this process."
 
     connection :steps, ProcessStepType.connection_type do
       description "All the steps of this process."
