+++ decidim-core/lib/decidim/core/api/session_type.rb
