+++ decidim-core/spec/types/translated_field_type_spec.rb
@@ -1,0 +1,67 @@
+# frozen_string_literal: true
+require "spec_helper"
+require "decidim/api/test/type_context"
+
+module Decidim
+  describe TranslatedFieldType do
+    include_context "graphql type"
+
+    let(:model) do
+      {
+        ca: "Hola",
+        en: "Hello"
+      }
+    end
+
+    describe "locales" do
+      let(:query) { "{ locales }" }
+
+      it "returns the available locales" do
+        expect(response["locales"]).to include("en", "ca")
+      end
+    end
+
+    describe "translations" do
+      context "when locales are not provided" do
+        let(:query) { '{ translations { locale text }}' }
+
+        it "returns all the translations" do
+          translations = response["translations"]
+          expect(translations.length).to eq(2)
+          expect(translations).to include({ "locale" => "ca", "text" => "Hola"})
+          expect(translations).to include({ "locale" => "en", "text" => "Hello"})
+        end
+      end
+
+      context "when locales are provided" do
+        let(:query) { '{ translations(locales: ["ca"]) { locale text }}' }
+
+        it "returns the translations on the provided locales" do
+          translations = response["translations"]
+          expect(translations).to include({ "locale" => "ca", "text" => "Hola"})
+          expect(translations).not_to include({ "locale" => "en", "text" => "Hello"})
+        end
+      end
+    end
+
+    describe "translation" do
+      context "when the locale is found" do
+        let(:query) { '{ translation(locale: "ca") }' }
+
+        it "returns the translation for that language" do
+          translation = response["translation"]
+          expect(translation).to eq("Hola")
+        end
+      end
+
+      context "when the locale is not found" do
+        let(:query) { '{ translation(locale: "fake") }' }
+
+        it "returns a null value" do
+          translation = response["translation"]
+          expect(translation).to eq(nil)
+        end
+      end
+    end
+  end
+end
